Python Bokeh for the Stock Info from Yahoo
With the Bokeh and yahoo_fin to get the stock information and show the stock line for the latest 90 days. This shows only the Zalando of Germany but it can show many lines together, and the code is the same except adding more line and change the show() function.

This part kann show the line of the stock as well as the detail information with mouse.

Hover part is for the dynamic show of the info and the source is transfer the data.

Yahoo_fin is very usefull to get the stock info of the whole world stocks. If you want to get the stock information and to analyse the stock with python, that is what you need.